docs: Clarify RNTL is tested to work with Jest (#98)
As discussed in https://github.com/callstack/react-native-testing-library/issues/96, I found that I could get RNTL working with Jest but not with Mocha. I haven't _confirmed_ it doesn't work with Mocha, but I think it would be helpful for users (like me) who don't have a strong preference which runner to use, to know that Jest could be a good first choice.