Add updateWithAct to render
To test behaviour that only occurs when props are updated it is
necessary to call update on the test instance. However, if the
behaviour under test is triggered by a hook, e.g. useEffect then
the test will fail because the hook is only triggered after the test
has run.

This addresses the issue by wrapping the update function with act().

https://github.com/callstack/react-native-testing-library/issues/111